+++ ContentPatcher/Framework/Migrations/Migration_1_25.cs
@@ -1,4 +1,9 @@
+using System;
 using System.Diagnostics.CodeAnalysis;
+using ContentPatcher.Framework.Conditions;
+using ContentPatcher.Framework.Constants;
+using ContentPatcher.Framework.Lexing.LexTokens;
+using ContentPatcher.Framework.Tokens;
 using StardewModdingAPI;
 
 namespace ContentPatcher.Framework.Migrations
@@ -13,5 +18,55 @@ namespace ContentPatcher.Framework.Migrations
         /// <summary>Construct an instance.</summary>
         public Migration_1_25()
             : base(new SemanticVersion(1, 25, 0)) { }
+
+        /// <inheritdoc />
+        public override bool TryMigrate(ref ILexToken lexToken, out string error)
+        {
+            if (!base.TryMigrate(ref lexToken, out error))
+                return false;
+
+            // 1.25 adds 'AnyPlayer' player type
+            if (lexToken is LexTokenToken token && Enum.TryParse(token.Name, ignoreCase: true, out ConditionType type) && this.IsPlayerToken(type) && token.HasInputArgs() && this.IsPlayerToken(type))
+            {
+                var input = new InputArguments(new LiteralString(token.InputArgs.ToString(), new LogPathBuilder()));
+                foreach (string value in input.PositionalArgs)
+                {
+                    if (Enum.TryParse(value, ignoreCase: true, out PlayerType playerType) && playerType is PlayerType.AnyPlayer)
+                    {
+                        error = this.GetNounPhraseError($"using the '{PlayerType.AnyPlayer}' player type");
+                        return false;
+                    }
+                }
+            }
+
+            return true;
+        }
+
+        /// <summary>Get whether a token type accepted a <see cref="PlayerType"/> input argument before Content Patcher 1.24.</summary>
+        /// <param name="type">The condition type.</param>
+        private bool IsPlayerToken(ConditionType type)
+        {
+            return type
+                is ConditionType.ChildGenders
+                or ConditionType.ChildNames
+                or ConditionType.DailyLuck
+                or ConditionType.FarmhouseUpgrade
+                or ConditionType.HasActiveQuest
+                or ConditionType.HasCaughtFish
+                or ConditionType.HasConversationTopic
+                or ConditionType.HasDialogueAnswer
+                or ConditionType.HasFlag
+                or ConditionType.HasProfession
+                or ConditionType.HasReadLetter
+                or ConditionType.HasSeenEvent
+                or ConditionType.IsMainPlayer
+                or ConditionType.IsOutdoors
+                or ConditionType.LocationContext
+                or ConditionType.LocationName
+                or ConditionType.LocationUniqueName
+                or ConditionType.PlayerGender
+                or ConditionType.PlayerName
+                or ConditionType.Spouse;
+        }
     }
 }